Refactor controllers to minimal apis.
- Refactor OtiumController
- Extract blocks to config
- Change auslastung to int percentage

+++ Afra-App/Data/DTO/Otium/LehrerUebersicht.cs
@@ -2,7 +2,7 @@
 
 public record LehrerUebersicht(IEnumerable<LehrerTerminPreview> termine, IEnumerable<MenteePreview> mentees);
 
-public record LehrerTerminPreview(Guid Id, string Otium, string Ort, double? Auslastung, DateOnly Datum, byte block);
+public record LehrerTerminPreview(Guid Id, string Otium, string Ort, int? Auslastung, DateOnly Datum, byte block);
 
 public record MenteePreview(PersonInfoMinimal mentee, MenteePreviewStatus letzteWoche, MenteePreviewStatus dieseWoche, MenteePreviewStatus nächsteWoche);
